+++ src/Cron/Mitgliederpruefung.php
@@ -28,6 +28,13 @@ class Mitgliederpruefung
 	 */
 	public function onHourly(): void
 	{
+
+		// ================================================================
+		// Prüfung tl_member auf Mitgliedschaft im BdF
+		// ================================================================
+		//$result = self::check_tl_member();
+		//return;
+		
 		//\System::getContainer()->get('monolog.logger.contao.cron')->info('[Fernschach-Wartung] yyy Mitglieder wurden überprüft');
 		// Alle Mitgliederkonten suchen, 
 		// 1) deren letzte Aktualisierung vor der letzten Wartung (fernschach_memberbridgeTime) erfolgte
@@ -198,6 +205,168 @@ class Mitgliederpruefung
 		// Log-Eintrag vornehmen
 		\System::getContainer()->get('monolog.logger.contao.cron')->info('[Fernschach-Wartung] Mitglieder wurden überprüft');
 		
+	}
+
+	/**
+	 * Funktion check_tl_member
+	 */
+	public function check_tl_member()
+	{
+		$objMember = \Database::getInstance()->prepare("SELECT * FROM tl_member WHERE locked = ? AND disable = ? AND username != ?")
+		                                     ->execute('', '', '');
+		$meldung = $objMember->numRows.' Frontend-Mitglieder müssen geprüft werden.<br>';
+
+		if($objMember->numRows)
+		{
+			// Alle Frontend-Mitglieder der Reihe nach prüfen
+			while($objMember->next())
+			{
+				// E-Mail-Adresse in Fernschach-Verwaltung suchen
+				$objPlayer = \Database::getInstance()->prepare("SELECT * FROM tl_fernschach_spieler WHERE (email1 = ? OR email2 = ?) AND published = ?")
+				                                     ->execute($objMember->email, $objMember->email, 1);
+
+				if($objPlayer->numRows)
+				{
+					// Gefundene Spieler prüfen
+					while($objPlayer->next())
+					{
+						// Ist der Spieler Mitglied im BdF?
+						$mitglied = \Schachbulle\ContaoFernschachBundle\Classes\Helper::checkMembership($objPlayer->memberships, NULL, $objPlayer->published);
+
+						if($mitglied)
+						{
+							// Spieler ist Mitglied und veröffentlicht
+							if($objMember->fernschach_memberId)
+							{
+								// Zuordnung bereits vorhanden, prüfen ob die zugeordnete ID paßt
+								if($objMember->fernschach_memberId == $objPlayer->id)
+								{
+									// ID's stimmen überein, jetzt Mitgliedergruppen prüfen
+									$gruppen = self::setGroups($objMember->groups, true); // Mitgliedergruppen aktualisieren, BdF-Mitglied eintragen
+									if($gruppen != $objMember->groups)
+									{
+										// Aktualisierung tl_member.groups notwendig
+										$set = array
+										(
+											'tstamp'                      => $aktuelleZeit,
+											'fernschach_memberbridgeTime' => $aktuelleZeit,
+											'groups'                      => $gruppen
+										);
+										\Database::getInstance()->prepare("UPDATE tl_member %s WHERE id=?")
+										                        ->set($set)
+										                        ->execute($objMember->id);
+										//$version = new \Versions('tl_member', $objMember->id);
+										//$version->setUsername($GLOBALS['TL_LANG']['fernschachverwaltung']['botname']);
+										//$version->create();
+
+										// Zuordnung entfernen
+										$meldung .= 'Zuordnung FE-Mitglied ('.$objMember->username.' - '.$objMember->firstname.' '.$objMember->lastname.') zu Gruppe BdF-Mitglied hinzugefügt.<br>';
+										//\System::log('[Fernschach-Wartung] Zuordnung FE-Mitglied ('.$objMember->username.' - '.$objMember->firstname.' '.$objMember->lastname.') zu Gruppe BdF-Mitglied hinzugefügt.', __CLASS__.'::'.__FUNCTION__, TL_GENERAL);
+									}
+								}
+								else
+								{
+									$gruppen = self::setGroups($objMember->groups, false); // Mitgliedergruppen aktualisieren, BdF-Mitglied austragen
+
+									// Datensatz aktualisieren
+									$set = array
+									(
+										'tstamp'                      => $aktuelleZeit,
+										'fernschach_memberbridgeTime' => $aktuelleZeit,
+										'groups'                      => $gruppen,
+										'fernschach_memberId'         => $objPlayer->id
+									);
+									\Database::getInstance()->prepare("UPDATE tl_member %s WHERE id=?")
+									                        ->set($set)
+									                        ->execute($objMember->id);
+									// Auskommentiert wegen ErrorException: Warning: Attempt to read property "server" on null in /kunden/107305_14053/webseiten/schachbund/dsbweb-entwicklung.2023/vendor/contao/core-bundle/src/Resources/contao/classes/Versions.php
+									//$version = new \Versions('tl_member', $objMember->id);
+									//$version->setUsername($GLOBALS['TL_LANG']['fernschachverwaltung']['botname']);
+									//$version->create();
+									//\System::log('[Fernschach-Wartung] tl_member.fernschach_memberId ('.$objMember->fernschach_memberId.') <> tl_fernschach_spieler.id ('.$objPlayer->id.'). Geändert von '.$objMember->fernschach_memberId.' auf '.$objPlayer->id.'.', __CLASS__.'::'.__FUNCTION__, TL_ERROR);
+								}
+							}
+							else
+							{
+								// Es gibt noch keine Zuordnung eines Mitglieds
+								$gruppen = self::setGroups($objMember->groups, true); // Mitgliedergruppen aktualisieren, BdF-Mitglied eintragen
+
+								// Datensatz aktualisieren
+								$set = array
+								(
+									'tstamp'                      => $aktuelleZeit,
+									'fernschach_memberbridgeTime' => $aktuelleZeit,
+									'groups'                      => $gruppen,
+									'fernschach_memberId'         => $objPlayer->id
+								);
+								\Database::getInstance()->prepare("UPDATE tl_member %s WHERE id=?")
+								                        ->set($set)
+								                        ->execute($objMember->id);
+								//$version = new \Versions('tl_member', $objMember->id);
+								//$version->setUsername($GLOBALS['TL_LANG']['fernschachverwaltung']['botname']);
+								//$version->create();
+
+								// Zuordnung noch nicht vorhanden, jetzt vornehmen
+								$meldung .= 'Neue Zuordnung FE-Mitglied ('.$objMember->username.' - '.$objMember->firstname.' '.$objMember->lastname.') zu BdF-Mitglied ('.$objPlayer->vorname.' '.$objPlayer->nachname.' ['.$objPlayer->id.']) vorgenommen.<br>';
+								//\System::log('[Fernschach-Wartung] Neue Zuordnung FE-Mitglied ('.$objMember->username.' - '.$objMember->firstname.' '.$objMember->lastname.') zu BdF-Mitglied ('.$objPlayer->vorname.' '.$objPlayer->nachname.' ['.$objPlayer->id.']) vorgenommen.', __CLASS__.'::'.__FUNCTION__, TL_GENERAL);
+							}
+						}
+						else
+						{
+							// Spieler ist kein Mitglied mehr, Zuordnung ggfs. entfernen
+							$gruppen = self::setGroups($objMember->groups, false); // Mitgliedergruppen aktualisieren, BdF-Mitglied austragen
+							if($objMember->fernschach_memberId > 0 || $gruppen != $objMember->groups)
+							{
+								// Aktualisierung tl_member notwendig
+								$set = array
+								(
+									'tstamp'                      => $aktuelleZeit,
+									'fernschach_memberbridgeTime' => $aktuelleZeit,
+									'groups'                      => $gruppen,
+									'fernschach_memberId'         => 0
+								);
+								\Database::getInstance()->prepare("UPDATE tl_member %s WHERE id=?")
+								                        ->set($set)
+								                        ->execute($objMember->id);
+								//$version = new \Versions('tl_member', $objMember->id);
+								//$version->setUsername($GLOBALS['TL_LANG']['fernschachverwaltung']['botname']);
+								//$version->create();
+
+								// Zuordnung entfernen
+								$meldung .= 'Zuordnung FE-Mitglied ('.$objMember->username.' - '.$objMember->firstname.' '.$objMember->lastname.') zu BdF-Mitglied entfernt, da ausgetreten.<br>';
+								//\System::log('[Fernschach-Wartung] Zuordnung FE-Mitglied ('.$objMember->username.' - '.$objMember->firstname.' '.$objMember->lastname.') zu BdF-Mitglied entfernt, da ausgetreten.', __CLASS__.'::'.__FUNCTION__, TL_GENERAL);
+							}
+						}
+					}
+				}
+				else
+				{
+					// Kein passender Spielerdatensatz gefunden, deshalb Zuordnung prüfen
+					$gruppen = self::setGroups($objMember->groups, false); // Mitgliedergruppen aktualisieren, BdF-Mitglied austragen
+					if($objMember->fernschach_memberId > 0 || $gruppen != $objMember->groups)
+					{
+						// Aktualisierung tl_member notwendig
+						$set = array
+						(
+							'tstamp'                      => $aktuelleZeit,
+							'fernschach_memberbridgeTime' => $aktuelleZeit,
+							'groups'                      => $gruppen,
+							'fernschach_memberId'         => 0
+						);
+						\Database::getInstance()->prepare("UPDATE tl_member %s WHERE id=?")
+						                        ->set($set)
+						                        ->execute($objMember->id);
+						//$version = new \Versions('tl_member', $objMember->id);
+						//$version->setUsername($GLOBALS['TL_LANG']['fernschachverwaltung']['botname']);
+						//$version->create();
+
+						// Zuordnung entfernen
+						$meldung .= 'Zuordnung FE-Mitglied ('.$objMember->username.' - '.$objMember->firstname.' '.$objMember->lastname.') zu BdF-Mitglied entfernt.';
+						//\System::log('[Fernschach-Wartung] Zuordnung FE-Mitglied ('.$objMember->username.' - '.$objMember->firstname.' '.$objMember->lastname.') zu BdF-Mitglied entfernt.', __CLASS__.'::'.__FUNCTION__, TL_GENERAL);
+					}
+				}
+			}
+		}
 	}
 
 	/**